Improve independent set of edges bound
The new bound combines the independent set of edges with the earlier
node degree sum bound. For the edges in the independent set, the maximum
degree of a contained node is summed. This value is subtracted from the
total number of edges, leaving the number of edges to hit. This number
is then covered by using degrees of available nodes.

This new bound is only used for pruning branches, forced nodes are still
using the older bound.